refactor: move components to a folder
