+++ client/src/index.js
@@ -10,6 +10,7 @@ ReactDOM.render(
     <BrowserRouter>
         <App/>
     </BrowserRouter>,
-    document.getElementById('root'));
+    document.getElementById('root')
+);
 
 unregisterServiceWorker();
